cleanup part1: the migration of SDL code from C64.cpp into mainsdl.cpp
